feat: add new utility for selected flyout component
