+++ class-components/my-app/src/components/selectedFlyout/SelectedFlyout.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useAppDispatch, useAppSelector } from '../../store/hooks';
 import { clearItems } from '../../store/selectedItemsSlice';
+import { downloadCSV } from '../../utils/downloadCSV';
 import './SelectedFlyout.css';
 
 const SelectedFlyout: React.FC = () => {
@@ -12,19 +13,7 @@ const SelectedFlyout: React.FC = () => {
   };
 
   const handleDownload = () => {
-    const csvRows = [
-      ['Name', 'URL'],
-      ...selectedItems.map((item) => [item.name, item.url]),
-    ];
-
-    const csvContent = csvRows.map((row) => row.join(',')).join('\n');
-    const blob = new Blob([csvContent], { type: 'text/csv' });
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = `${selectedItems.length}_items.csv`;
-    a.click();
-    URL.revokeObjectURL(url);
+    downloadCSV(selectedItems);
   };
 
   if (selectedItems.length === 0) return null;
